add forgot password

+++ src/Controller/SecurityController.php
@@ -4,6 +4,8 @@ namespace App\Controller;
 
 use App\Entity\VerificationEmail;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -22,29 +24,25 @@ class SecurityController extends Controller
     public function registerAction(Request $request, UserPasswordEncoderInterface $passwordEncoder, \Swift_Mailer $mailer)
     {
         // 1) build the form
-        $verificatedSting = $this->random_str("alphanum", 64);
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         // 2) handle the submit (will only happen on POST)
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
+            $verificatedSting = $this->random_str("alphanum", 64);
             // 3) Encode the password (you could also do this via Doctrine listener)
             $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
             $user->setPassword($password);
 
-            //random str to verification user
             $verification = new VerificationEmail();
             $verification->setUser($user);
             $verification->setVerificationString($verificatedSting);
 
-            $this->sendEmail($verificatedSting, $user, $mailer);
-            // 4) save the User!
+            $this->sendEmail($verificatedSting, $user, $mailer, 'verif');
             $em = $this->getDoctrine()->getManager();
             $em->persist($user);
             $em->persist($verification);
             $em->flush();
-            // ... do any other work - like sending them an email, etc
-            // maybe set a "flash" success message for the user
             return $this->redirectToRoute('homepage');
         }
 
@@ -55,52 +53,43 @@ class SecurityController extends Controller
     }
 
     /**
-     * @Route("/email_verification/{id}", name="verification")
+     * @Route("/email_verification/{id}", name="email_verif")
      *
      */
-    public function emailVerification(Request $request, $id)
+    public function emailVerification($id)
     {
         $em = $this->getDoctrine()->getManager();
         $email = $em->getRepository(VerificationEmail::class)->findOneBy(['verificationString' => $id]);
-        $userEmail = $email->getUser();
-        $em->getRepository(User::class)->find($userEmail)->setRoles(array("ROLE_USER"));
-        $em->remove($email);
-
-        $em->flush();
+        if (!$email->getForgot()) {
+            $userEmail = $email->getUser();
+            $em->getRepository(User::class)->find($userEmail)->setRoles(array("ROLE_USER"));
+            $em->remove($email);
+            $em->flush();
+        }
         return $this->redirectToRoute("homepage");
     }
 
     /**
      * @Route("/user_edit/{id}", name="user_edit")
-     *
+     * @Security("has_role('ROLE_ADMIN')")
      */
     public function editAction(Request $request, UserPasswordEncoderInterface $passwordEncoder, $id)
     {
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository(User::class)->find($id);
-
         if (!$user) {
             throw $this->createNotFoundException(
                 'No product found for id ' . $id
             );
         }
-
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
-            // 3) Encode the password (you could also do this via Doctrine listener)
             $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
             $user->setPassword($password);
-
-            // 4) save the User
             $em->flush();
-            // ... do any other work - like sending them an email, etc
-            // maybe set a "flash" success message for the user
-
             return $this->redirectToRoute('homepage');
         }
-
         return $this->render(
             'security/registration.html.twig',
             array('form' => $form->createView())
@@ -108,10 +97,10 @@ class SecurityController extends Controller
     }
 
     /**
-     * @Route("/user_delete", name="user_delete"):void
+     * @Route("/user_delete", name="user_delete")
      * @Security("has_role('ROLE_ADMIN')")
      */
-    public function user_delete_Action(Request $request,AuthorizationCheckerInterface $authChecker)
+    public function user_delete_Action(Request $request)
     {
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository(User::class);
@@ -120,6 +109,61 @@ class SecurityController extends Controller
         }
         $em->flush();
         return $this->redirectToRoute("homepage");
+    }
+
+    /**
+     * @Route("/forgot", name="forgot")
+     */
+    public function forgotPassword(Request $request, \Swift_Mailer $mailer)
+    {
+        $email = $request->query->get("fp_email");
+        if ($email) {
+            $verificatedSting = $this->random_str("alphanum", 64);
+
+            $em = $this->getDoctrine()->getManager();
+            $user = $em->getRepository(User::class)->findOneBy(['email' => $email]);
+
+            $verification = new VerificationEmail();
+
+            $verification->setUser($user);
+            $verification->setVerificationString($verificatedSting);
+            $verification->setForgot(true);
+
+            $em->persist($verification);
+            $em->flush();
+
+            $this->sendEmail($verificatedSting, $user, $mailer, 'forgot');
+        }
+        return $this->render('security/forgot.html.twig');
+    }
+
+    /**
+     * @Route("/forgot_passowrd/{id}")
+     */
+    public function forgotPasswordAction(Request $request, UserPasswordEncoderInterface $passwordEncoder, $id)
+    {
+        $em = $this->getDoctrine()->getManager();
+        $email = $em->getRepository(VerificationEmail::class)->findOneBy(['verificationString' => $id]);
+
+        if ($email->getForgot()) {
+            $user = $em->getRepository(User::class)->find($email->getUser());
+            $form = $this->createFormBuilder($user)->add('plainPassword', RepeatedType::class, array(
+                'type' => PasswordType::class,
+                'first_options' => array('label' => 'Password'),
+                'second_options' => array('label' => 'Repeat Password'),
+            ))->getForm();
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                // 3) Encode the password (you could also do this via Doctrine listener)
+                $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
+                $user->setPassword($password);
+                $em->remove($email);
+                $em->flush();
+                return $this->redirectToRoute('homepage');
+            }
+        }
+        return $this->render("security/forgotedit.html.twig",
+            array('form' => $form->createView()));
     }
 
     function random_str($type = 'alphanum', $length = 8)
@@ -153,16 +197,35 @@ class SecurityController extends Controller
         }
     }
 
-    public function sendEmail(string $str, User $user, \Swift_Mailer $mailer): void
+    public function sendEmail(string $str, User $user, \Swift_Mailer $mailer, $type): void
     {
-        $message = (new \Swift_Message('Verification user'))
-            ->setFrom('itrasymfony@gmail.com')
-            ->setTo($user->getEmail())
-            ->setBody(
-                new Response("http:/symfony.dev/email_verification/" . $str
-                ),
-                'text/html'
-            );
+        switch ($type) {
+
+            case 'verif':
+                {
+                    $message = (new \Swift_Message('Verification user'))
+                        ->setFrom('itrasymfony@gmail.com')
+                        ->setTo($user->getEmail())
+                        ->setBody(
+                            new Response("http:/symfony.dev/email_verification/" . $str
+                            ),
+                            'text/html'
+                        );
+                    break;
+                }
+            case 'forgot':
+                {
+                    $message = (new \Swift_Message('Forgot password'))
+                        ->setFrom('itrasymfony@gmail.com')
+                        ->setTo($user->getEmail())
+                        ->setBody(
+                            new Response("http:/symfony.dev/forgot_passowrd/" . $str
+                            ),
+                            'text/html'
+                        );
+                    break;
+                }
+        }
         $mailer->send($message);
         return;
     }
